The history route and his functions are been created
